requirements.txt
evdev
pygame
requests
python-dotenv
boto3
openai
beautifulsoup4
flask
